+++ tests/unit/plugins/modules/test_minio_policy.py
@@ -8,7 +8,13 @@ from __future__ import absolute_import, division, print_function
 
 __metaclass__ = type
 
-from unittest.mock import patch, ANY, MagicMock
+import pytest
+
+from ansible_collections.dubzland.minio.tests.unit.compat.mock import (
+    patch,
+    ANY,
+    MagicMock,
+)
 
 from ansible_collections.dubzland.minio.tests.unit.plugins.modules.utils import (
     AnsibleExitJson,
@@ -17,21 +23,23 @@ from ansible_collections.dubzland.minio.tests.unit.plugins.modules.utils import 
     set_module_args,
 )
 
-from ansible_collections.dubzland.minio.plugins.modules import minio_admin_policy
+from ansible_collections.dubzland.minio.plugins.modules import minio_policy
 
 
-class TestMinioAdminPolicy(ModuleTestCase):
+class TestMinioPolicy(ModuleTestCase):
     def setUp(self):
-        super().setUp()
+        super(TestMinioPolicy, self).setUp()
+
+    @pytest.fixture(autouse=True)
+    def _mocker(self, mocker):
+        self.mocker = mocker
 
     def test_module_fail_when_args_missing(self):
         with self.assertRaises(AnsibleFailJson) as r:
             set_module_args({})
-            minio_admin_policy.main()
+            minio_policy.main()
 
-    @patch(
-        "ansible_collections.dubzland.minio.plugins.modules.minio_admin_policy.MinioAdmin"
-    )
+    @patch("ansible_collections.dubzland.minio.plugins.modules.minio_policy.MinioAdmin")
     def test_module_unchanged_when_exists(self, mock_admin_client):
         client = MagicMock()
         client.policy_list.return_value = '{"%s": {}}' % "testing"
@@ -47,11 +55,9 @@ class TestMinioAdminPolicy(ModuleTestCase):
                     "minio_url": "http://localhost:9000",
                 }
             )
-            minio_admin_policy.main()
+            minio_policy.main()
 
-    @patch(
-        "ansible_collections.dubzland.minio.plugins.modules.minio_admin_policy.MinioAdmin"
-    )
+    @patch("ansible_collections.dubzland.minio.plugins.modules.minio_policy.MinioAdmin")
     def test_module_update_when_different(self, mock_admin_client):
         client = MagicMock()
         client.policy_list.return_value = '{"%s": {"existing": "value"}}'
@@ -67,11 +73,9 @@ class TestMinioAdminPolicy(ModuleTestCase):
                     "minio_url": "http://localhost:9000",
                 }
             )
-            minio_admin_policy.main()
+            minio_policy.main()
 
-    @patch(
-        "ansible_collections.dubzland.minio.plugins.modules.minio_admin_policy.MinioAdmin"
-    )
+    @patch("ansible_collections.dubzland.minio.plugins.modules.minio_policy.MinioAdmin")
     def test_module_create_when_not_exist(self, mock_admin_client):
         client = MagicMock()
         client.policy_list.return_value = "{}"
@@ -88,16 +92,14 @@ class TestMinioAdminPolicy(ModuleTestCase):
                     "minio_url": "http://localhost:9000",
                 }
             )
-            minio_admin_policy.main()
+            minio_policy.main()
 
         client.policy_list.assert_called_once()
         client.policy_add.assert_called_with("testing", ANY)
         result = r.exception.args[0]
         assert result["changed"] is True
 
-    @patch(
-        "ansible_collections.dubzland.minio.plugins.modules.minio_admin_policy.MinioAdmin"
-    )
+    @patch("ansible_collections.dubzland.minio.plugins.modules.minio_policy.MinioAdmin")
     def test_module_unchanged_when_not_exist_and_state_absent(self, mock_admin_client):
         client = MagicMock()
         client.policy_list.return_value = "{}"
@@ -115,16 +117,14 @@ class TestMinioAdminPolicy(ModuleTestCase):
                     "state": "absent",
                 }
             )
-            minio_admin_policy.main()
+            minio_policy.main()
 
         client.policy_list.assert_called_once()
         client.policy_add.assert_not_called()
         result = r.exception.args[0]
         assert result["changed"] is False
 
-    @patch(
-        "ansible_collections.dubzland.minio.plugins.modules.minio_admin_policy.MinioAdmin"
-    )
+    @patch("ansible_collections.dubzland.minio.plugins.modules.minio_policy.MinioAdmin")
     def test_module_changed_when_exists_and_state_absent(self, mock_admin_client):
         client = MagicMock()
         client.policy_list.return_value = '{"testing": {}}'
@@ -142,7 +142,7 @@ class TestMinioAdminPolicy(ModuleTestCase):
                     "state": "absent",
                 }
             )
-            minio_admin_policy.main()
+            minio_policy.main()
 
         client.policy_list.assert_called_once()
         client.policy_remove.assert_called_once_with("testing")