+++ tests/unit/plugins/modules/utils.py
@@ -9,13 +9,12 @@ from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
 import json
-import unittest
 
-from contextlib import contextmanager
-
-from unittest.mock import patch
+from ansible_collections.dubzland.minio.tests.unit.compat import unittest
+from ansible_collections.dubzland.minio.tests.unit.compat.mock import patch
 from ansible.module_utils import basic
 from ansible.module_utils.common.text.converters import to_bytes
+from contextlib import contextmanager
 
 
 class AnsibleExitJson(Exception):